+++ app/Http/Controllers/HandleidingController.php
@@ -81,7 +81,8 @@ class HandleidingController extends Controller
             ['nr' => 17, 'titel' => 'Stichtingsbestuur', 'rollen' => ['stichtingsbestuur']],
             ['nr' => 18, 'titel' => 'Noodtoegang', 'rollen' => ['beheerder']],
             ['nr' => 19, 'titel' => 'Quote in de zijbalk', 'rollen' => []],
-            ['nr' => 20, 'titel' => 'Vragen of problemen?', 'rollen' => []],
+            ['nr' => 20, 'titel' => 'Systeemmeldingen', 'rollen' => ['beheerder']],
+            ['nr' => 21, 'titel' => 'Vragen of problemen?', 'rollen' => []],
         ];
     }
 }